server/websocket-service/requirements.txt
fastapi>=0.115.12
python-socketio==5.13.0
python-jose==3.4.0
httpx>=0.28.1
uvicorn>=0.34.0
websockets==15.0.1
python-dotenv>=1.1.0
redis==5.2.1
